requirement.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
supabase
httpx
pypdf
python-docx